Renamed class Request to genConfig
