+++ test/cli/_gen2/test_callback.py
@@ -1,15 +1,15 @@
 import unittest
-from xcube.cli._gen2.request import Callback
+from xcube.cli._gen2.genconfig import CallbackConfig
 
 
-class TestCallback(unittest.TestCase):
+class TestCallbackConfig(unittest.TestCase):
     def test_callback(self):
         with self.assertRaises(ValueError) as e:
-            Callback()
+            CallbackConfig()
         self.assertEqual('Both, api_uri and access_token must be given', str(e.exception))
 
         expected = {"api_uri": 'https://bla.com', "access_token": 'dfasovjdasoövjidfs'}
-        callback = Callback(**expected)
+        callback = CallbackConfig(**expected)
         res = callback.to_dict()
 
         self.assertDictEqual(expected, res)
